Restructure the code to be a class.
